+++ moduleMap.js
@@ -1,0 +1,20 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @format
+ */
+'use strict';
+
+const moduleMap = Object.assign(
+  {
+    React: 'react',
+    ReactDOM: 'react-dom',
+  },
+  require('fbjs/module-map'),
+  require('fbjs-scripts/third-party-module-map'),
+);
+
+module.exports = moduleMap;